refactor: routes, middlewares and services
