+++ app/backend/src/routes/taskRoutes.js
@@ -1,0 +1,8 @@
+const express = require('express');
+const { getAllTasks } = require('../controllers/taskController');
+
+const router = express.Router();
+
+router.get('/', getAllTasks);
+
+module.exports = router;
