+++ app/backend/src/services/taskService.js
@@ -1,0 +1,11 @@
+const { Task } = require('../models');
+
+const getAllTasks = async () => {
+  const tasks = await Task.findAll();
+
+  return tasks;
+};
+
+module.exports = {
+  getAllTasks,
+};
